separated IT tests
